+++ src/ManagedModels/Actions/Duplicate/DuplicateFragment.php
@@ -9,6 +9,7 @@ use Thinktomorrow\Chief\Fragments\Database\ContextModel;
 use Thinktomorrow\Chief\Fragments\Database\FragmentModel;
 use Thinktomorrow\Chief\Fragments\Database\FragmentRepository;
 use Thinktomorrow\Chief\Fragments\Events\FragmentDuplicated;
+use Thinktomorrow\Chief\Fragments\Exceptions\FragmentAlreadyAdded;
 
 class DuplicateFragment
 {
@@ -31,30 +32,32 @@ class DuplicateFragment
      * - a nested dynamic fragment is not duplicated but shared (e.g. article, quote)
      *
      * @param Model $targetModel
-     * @param FragmentModel $fragment
+     * @param FragmentModel $fragmentModel
      * @param int $index
+     * @param bool $forceDuplicateDynamicFragment
      * @param int $level
-     * @throws \Thinktomorrow\Chief\Fragments\Exceptions\FragmentAlreadyAdded
+     * @throws FragmentAlreadyAdded
      */
-    public function handle(Model $targetModel, FragmentModel $fragment, int $index, $forceDuplicateSharedFragment = false, $level = 0): void
+    public function handle(Model $targetModel, FragmentModel $fragmentModel, int $index, bool $forceDuplicateDynamicFragment = false, $level = 0): void
     {
         if (! $contextModel = ContextModel::ownedBy($targetModel)) {
             $contextModel = ContextModel::createForOwner($targetModel);
         }
 
         // If it's a shareable fragment, we'll use the original
-        if (! $forceDuplicateSharedFragment && ($fragment->isShared() || ($level > 0 && ! $fragment->refersToStaticObject()))) {
-            $this->addFragmentModel->handle($targetModel, $fragment, $index);
+        if (! $forceDuplicateDynamicFragment && ! $fragmentModel->refersToStaticObject()) {
+            $this->addFragmentModel->handle($targetModel, $fragmentModel, $index);
 
             return;
         }
 
         // Otherwise do a full copy of the fragment instead
-        $copiedFragment = $fragment->replicate();
+        $copiedFragment = $fragmentModel->replicate();
         $copiedFragment->id = $this->fragmentRepository->nextId();
 
+        // TODO: still to check if this is desired that we copy the dynamic model as well...
         if ($copiedFragment->refersToDynamicModel()) {
-            $model = $this->duplicateModel->handle($this->fragmentRepository->find($fragment->id));
+            $model = $this->duplicateModel->handle($this->fragmentRepository->find($fragmentModel->id));
             $copiedFragment->model_reference = $model->modelReference()->getShort();
         }
 
@@ -62,18 +65,15 @@ class DuplicateFragment
 
         $contextModel->fragments()->attach($copiedFragment, ['order' => $index]);
 
-        // TODO: keep same order...
-
-        // Assets
-//        foreach ($fragment->assets() as $asset) {
-//            $copiedFragment->assetRelation()->attach($asset, ['type' => $asset->pivot->type, 'locale' => $asset->pivot->locale, 'order' => $asset->pivot->order]);
-//        }
+        foreach ($fragmentModel->assets() as $asset) {
+            $copiedFragment->assetRelation()->attach($asset, ['type' => $asset->pivot->type, 'locale' => $asset->pivot->locale, 'order' => $asset->pivot->order]);
+        }
 
         event(new FragmentDuplicated($copiedFragment->id, $contextModel->id));
 
         // Handle nested fragments
 
-        foreach ($this->fragmentRepository->getByOwner($fragment) as $i => $nestedFragment) {
+        foreach ($this->fragmentRepository->getByOwner($fragmentModel) as $i => $nestedFragment) {
             $this->handle($copiedFragment, $nestedFragment->fragmentModel(), $i, false, ++$level);
         }
     }