+++ tests/Application/Fragments/CopyFragmentTest.php
@@ -6,7 +6,7 @@ use Thinktomorrow\Chief\Tests\ChiefTestCase;
 use Thinktomorrow\Chief\Tests\Shared\Fakes\ArticlePage;
 use Thinktomorrow\Chief\Tests\Shared\Fakes\Quote;
 
-class DuplicateFragmentTest extends ChiefTestCase
+class CopyFragmentTest extends ChiefTestCase
 {
     private ArticlePage $owner;
 
@@ -18,7 +18,7 @@ class DuplicateFragmentTest extends ChiefTestCase
     }
 
     /** @test */
-    public function it_can_duplicate_a_static_fragment()
+    public function it_can_copy_a_static_fragment()
     {
         $snippet = $this->setupAndCreateSnippet($this->owner);
         $newOwner = ArticlePage::create();
@@ -35,7 +35,7 @@ class DuplicateFragmentTest extends ChiefTestCase
     }
 
     /** @test */
-    public function a_dynamic_fragment_can_be_duplicated()
+    public function by_default_a_dynamic_fragment_is_copied_as_reference()
     {
         $quote = $this->setupAndCreateQuote($this->owner);
         $newOwner = ArticlePage::create();
@@ -44,18 +44,20 @@ class DuplicateFragmentTest extends ChiefTestCase
 
         $this->assertFragmentCount($newOwner, 1);
 
-        $this->assertCount(2, Quote::all());
+        $this->assertCount(1, Quote::all());
 
         $existingFragmentFresh = $this->firstFragment($this->owner);
         $fragmentFresh = $this->firstFragment($newOwner);
 
-        $this->assertNotEquals($existingFragmentFresh->fragmentModel()->id, $fragmentFresh->fragmentModel()->id);
-        $this->assertNotEquals($existingFragmentFresh->fragmentModel()->model_reference, $fragmentFresh->fragmentModel()->model_reference);
-        $this->assertNotEquals($existingFragmentFresh->id, $fragmentFresh->id);
+        $this->assertEquals($existingFragmentFresh->id, $fragmentFresh->id);
+        $this->assertEquals($existingFragmentFresh->fragmentModel()->id, $fragmentFresh->fragmentModel()->id);
+        $this->assertEquals($existingFragmentFresh->fragmentModel()->model_reference, $fragmentFresh->fragmentModel()->model_reference);
+        $this->assertTrue($existingFragmentFresh->fragmentModel()->isShared());
+        $this->assertTrue($fragmentFresh->fragmentModel()->isShared());
     }
 
     /** @test */
-    public function static_nested_fragments_are_duplicated_as_well()
+    public function static_nested_fragments_are_copied_when_fragment_is_copied()
     {
         $quote = $this->setupAndCreateQuote($this->owner);
         $newOwner = ArticlePage::create();
@@ -63,7 +65,9 @@ class DuplicateFragmentTest extends ChiefTestCase
 
         $this->createAsFragment($nestedFragment, $quote->fragmentModel());
 
-        $this->asAdmin()->post($this->manager($quote)->route('fragment-copy', $newOwner, $quote));
+        $this->asAdmin()->post($this->manager($quote)->route('fragment-copy', $newOwner, $quote), [
+            'hardcopy' => true,
+        ]);
 
         $this->assertFragmentCount($newOwner, 1);
 
@@ -96,22 +100,5 @@ class DuplicateFragmentTest extends ChiefTestCase
 
         $this->assertEquals($existingNestedFragmentFresh->id, $nestedFragmentFresh->id);
         $this->assertEquals($existingNestedFragmentFresh->fragmentModel()->id, $nestedFragmentFresh->fragmentModel()->id);
-    }
-
-    /** @test */
-    public function a_nested_shared_fragment_is_kept_shared()
-    {
-        $quote = $this->setupAndCreateQuote($this->owner);
-
-        $newOwner = ArticlePage::create();
-
-        $nestedFragment = $this->createAsFragment(ArticlePage::create(), $quote->fragmentModel());
-
-        $this->asAdmin()->post($this->manager($quote)->route('fragment-copy', $newOwner, $quote));
-
-        $fragmentFresh = $this->firstFragment($newOwner);
-        $nestedFragmentFresh = $this->firstFragment($fragmentFresh->fragmentModel());
-
-        $this->assertTrue($nestedFragmentFresh->fragmentModel()->isShared());
     }
 }